backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
SQLAlchemy>=2.0.0
asyncpg>=0.27.0
alembic>=1.10.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
celery>=5.3.0
redis>=4.5.0
psycopg2-binary>=2.9.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
argon2-cffi>=21.3.0
python-multipart>=0.0.6
email-validator>=2.0.0
pandas>=2.3.3
openpyxl>=3.1.5
aiohttp>=3.13.2
feedparser>=6.0.12
tqdm>=4.67.1

aiohappyeyeballs==2.6.1
aiosignal==1.4.0
alabaster==1.0.0
anyio==4.11.0
arrow==1.4.0
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
bibtexparser==1.4.3
certifi==2025.11.12
charset-normalizer==3.4.4
Deprecated==1.3.1
docutils==0.21.2
fake-useragent==2.2.0
free_proxy==1.1.3
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imagesize==1.4.1
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
multidict==6.7.0
outcome==1.3.0.post0
packaging==25.0
propcache==0.4.1
Pygments==2.19.2
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
requests==2.32.5
roman-numerals-py==3.1.0
scholarly==1.7.11
selenium==4.38.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
soupsieve==2.8
Sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.9.0
wrapt==2.0.1
wsproto==1.3.2
yarl==1.22.0
